python_back_end/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow 